To handle use of "subject_id" as a field name
Some projects use "subject_id" as a valid field name in their subject metadata. This conflicts with the use of subject_id as a replacement name for subject_ids when the nfn column names are renamed and sorted.  
Also simplify conversion of subject_ids to subject_id as a column header name using rename() rather than map()